Fix page fault in rpool allocation to offline CPUs
In cases where HT is enabled in BIOS but disabled in OS
we run into a situation where we are allocating resources
and attempting to schedule work on offline cores which causes
a page fault and resets the machine.

This change checks the number of possible CPUs and checks
that each is online before allocating rpools. Fixes issue #165

Signed-off-by: Nathan Ciobanu <nathan.d.ciobanu@intel.com>