Return error on cache mode change in standby mode
Changing cache mode to the same mode is a special case that in OCL is
handled on the kernel level, without calling an OCF API. In result it
seemed to succeed even in standby mode, which should return an error.

Explicitly check for standby to return an appropriate error code.


Signed-off-by: Robert Baldyga <robert.baldyga@intel.com>