packaging: Fix debug package build
For debug symbols packages to build properly, we need to
instruct the compiler to produce debugging information
during the compilation process by adding a proper flag.
Additionally there is no point to create packages with
debug info in normal build, because it may crash the
package creation process if no debug info is found.

Signed-off-by: Rafal Stefanowski <rafal.stefanowski@intel.com>