cas_disk: adapt to kernel 5.13
Kernel commit c76f48eb add a lock of bd_mutex in the del_gendisk
function. This cause a deadlock in cas_disk as it hold this lock
before calling del_gendisk.

This patch fix this issue by holding the bd_mutex only when
accessing the bd_openers field and releasing it afterward. It
also add an open ops function which prevents an access to the
disk while it is shutting down (locked_bd is set).

Signed-off-by: Gal Hammer <gal.hammer@huawei.com>
Signed-off-by: Shai Fultheim <shai.fultheim@huawei.com>