Fix test_cleaning_policy_config ALRU assumptions
ALRU does not have a configurable sleep time between active awakenings
of the cleaning thread. Once it starts cleaning, it does not back off
unless staleness time and/or activity threshold constraints force it
to do so.

Changing test_cleaning_policy_config assumptions so that it only checks
whether minimum expected amount of data is cleaned rather than assume
exact amount of data to be cleaned.

Signed-off-by: Adam Rutkowski <adam.j.rutkowski@intel.com>