tests: remove test_trim_eviction
There are multiple issues with this test, but regardless
of them it seems that it duplicates existing trim tests.

The intention of the test is to verify that there is no
eviction when inserting to cache after a trim. This
is equivalent to checking that a trim reduced cache
occupancy, and this is done in other trim test
(test_trim_device_discard_support).

Signed-off-by: Rutkowski, Adam J <adam.j.rutkowski@intel.com>